+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   * Update analyzer dependency to `5.2.0`.
   * Address analyzer deprecations.
   * Require Dart 2.18.
+  * Fix issue #182, enabling escaping of curly brackets with a single quote as per the ICU message format specs.
 
 ## 0.18.0
   * Add support for Flutter locale split.